requirements.txt
tensorflow
librosa
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
pyyaml